[Simulator] Refactoring Utilities into utils.py.
